Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
email-validator>=2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.17
openpyxl==3.1.5
anthropic==0.39.0
httpx==0.27.2
PyJWT==2.11.0
